backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
casparser>=0.7.3
python-multipart>=0.0.6
pydantic>=2.5.3
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
